requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.7
django-environ==0.7.0
djangorestframework==3.12.4
gunicorn==20.1.0
kafka-python==2.0.2
pika==1.2.0
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
